requirements.txt
Pillow
numpy
matplotlib
mtcnn
tensorflow
deepface
